파일구조 수정 및 인터페이스 정의
파일의 전체 구조를 변경하고 각 인터페이스를 정의한 기능에 따라 정의하였음
