--- src/main/java/baseball/game/GameSetting.java
@@ -1,4 +1,0 @@
-package baseball.game;
-
-public interface GameSetting {
-}
