--- src/main/java/baseball/game/Hitter.java
@@ -1,4 +1,0 @@
-package baseball.game;
-
-public interface Hitter {
-}
