--- src/main/java/baseball/game/Pitcher.java
@@ -1,4 +1,0 @@
-package baseball.game;
-
-public interface Pitcher {
-}
