+++ src/main/java/baseball/game/Score.java
@@ -1,4 +1,7 @@
 package baseball.game;
 
 public class Score {
+
+    private int strikeCount;
+    private int ballCount;
 }
