--- src/main/java/baseball/game/Umpire.java
@@ -1,4 +1,0 @@
-package baseball.game;
-
-public interface Umpire {
-}
