+++ src/main/java/baseball/game/gamesetter/GameSetting.java
@@ -1,0 +1,8 @@
+package baseball.game.gamesetter;
+
+public interface GameSetting {
+
+    void initGame();
+    void restartGame();
+    void setGame();
+}
