+++ src/main/java/baseball/game/gamesetter/GameSettingImpl.java
@@ -1,0 +1,6 @@
+package baseball.game.gamesetter;
+
+import baseball.game.gamesetter.GameSetting;
+
+public class GameSettingImpl implements GameSetting {
+}
