+++ src/main/java/baseball/game/hitter/Hitter.java
@@ -1,0 +1,5 @@
+package baseball.game.hitter;
+
+public interface Hitter {
+    String swing();
+}
