+++ src/main/java/baseball/game/hitter/PlayerHitterImpl.java
@@ -1,0 +1,4 @@
+package baseball.game.hitter;
+
+public class PlayerHitterImpl implements Hitter {
+}
