--- src/main/java/baseball/game/impl/ComputerPitcherImpl.java
@@ -1,6 +1,0 @@
-package baseball.game.impl;
-
-import baseball.game.Pitcher;
-
-public class ComputerPitcherImpl implements Pitcher {
-}
