--- src/main/java/baseball/game/impl/ComputerUmpireImpl.java
@@ -1,6 +1,0 @@
-package baseball.game.impl;
-
-import baseball.game.Umpire;
-
-public class ComputerUmpireImpl implements Umpire {
-}
