--- src/main/java/baseball/game/impl/GameSettingImpl.java
@@ -1,6 +1,0 @@
-package baseball.game.impl;
-
-import baseball.game.GameSetting;
-
-public class GameSettingImpl implements GameSetting {
-}
