--- src/main/java/baseball/game/impl/PlayerHitterImpl.java
@@ -1,6 +1,0 @@
-package baseball.game.impl;
-
-import baseball.game.Hitter;
-
-public class PlayerHitterImpl implements Hitter {
-}
