+++ src/main/java/baseball/game/pitcher/ComputerPitcherImpl.java
@@ -1,0 +1,4 @@
+package baseball.game.pitcher;
+
+public class ComputerPitcherImpl implements Pitcher {
+}
