+++ src/main/java/baseball/game/pitcher/Pitcher.java
@@ -1,0 +1,5 @@
+package baseball.game.pitcher;
+
+public interface Pitcher {
+    int pitch();
+}
