+++ src/main/java/baseball/game/umpire/ComputerUmpireImpl.java
@@ -1,0 +1,4 @@
+package baseball.game.umpire;
+
+public class ComputerUmpireImpl implements Umpire {
+}
